ai-service/requirements.txt
# AI Service - Requirements
# Smart AI Engineering Platform - SPKBG

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# YOLO & Computer Vision
ultralytics==8.1.2
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3

# Utilities
pydantic==2.5.3
Pillow==10.2.0
requests==2.31.0

# Development (optional)
pytest==7.4.4
httpx==0.26.0
